+++ app/Http/Controllers/Api/AuthController.php
@@ -10,25 +10,24 @@ use App\Http\Requests\RegisterRequest;
 use App\Http\Resources\UserResource;
 use App\Interfaces\AuthenticationInterface;
 use App\Models\User;
+use App\Services\AuthServices;
 
 class AuthController extends Controller
 {
-    public function __construct(private AuthenticationInterface $authenticationInterface)
+    public function __construct(private AuthServices $authServices)
     {
         
     }
 
     public function register(RegisterRequest $request)
     {
-        $user = User::create($request->validated());
-        $token = $this->authenticationInterface->getToken($user, 'auth_token');
+        list($user, $token) = $this->authServices->register($request);
         return ApiResponse::sendResponse(new UserResource($user), 'User created!', 201, $token);
     }
 
     public function login(LoginRequest $request)
     {
-        $user = User::where('email', $request->email)->first();
-        $token = $this->authenticationInterface->validateCredentials($request, $user, 'auth_token');
+        list($user, $token) = $this->authServices->login($request);
         return ApiResponse::sendResponse(new UserResource($user), 'User logged in!', 200, $token);
     }
 }
